Items UX: Change flow for new items, add editing feature
- Adds a form for encapsulation reasons, add editing an item
- Adding a new item is now a quick process: Enter the name, save item
and directly go into edit item mode. For reason, see next point.
- Item Category: is saved directly after changing it, in order to
properly display the change in the vue-multiselect. vue-multiselect
only displays the name properly if the item.category is passed in as a
value. But since this a vuex-orm related field only populated while
querying the instance, saving the item instance directly is the
solution. Otherwise the previous value would continue to be displayed.